tsparticles and start break games and settings into components
